+++ src/directive/angular-tree-dnd-nodes.js
@@ -1,21 +1,12 @@
 angular.module('ntt.TreeDnD')
     .directive(
-    'treeDndNodes', function () {
+    'treeDndNodes', ['$parse',function ($parse) {
         return {
             restrict: 'A',
             replace:  true,
             link:     function (scope, element, attrs) {
-                scope.nodes = [];
                 scope.$nodes_class = '';
                 scope.$type = 'TreeDnDNodes';
-                scope.$element = element;
-                scope.getScopeNode = null;
-
-                scope.$watch(
-                    attrs.treeDndNodes, function (newValue, oldValue, scope) {
-                        scope.nodes = newValue;
-                    }, true
-                );
 
                 if (scope.$class.nodes) {
                     element.addClass(scope.$class.nodes);
@@ -23,5 +14,5 @@ angular.module('ntt.TreeDnD')
                 }
             }
         };
-    }
+    }]
 );